+++ src/rendering/buffer/buffer_helper.cpp
@@ -111,11 +111,11 @@ void copyResource(ID3D12GraphicsCommandList* cmdList,
 void copyBufferRegion(ID3D12GraphicsCommandList* cmdList,
                       ID3D12Resource* destBuffer,
                       D3D12_RESOURCE_STATES destState,
-                      uint32_t destOffsetBytes,
+                      size_t destOffsetBytes,
                       ID3D12Resource* srcBuffer,
                       D3D12_RESOURCE_STATES srcState,
-                      uint32_t srcOffsetBytes,
-                      uint32_t sizeBytes)
+                      size_t srcOffsetBytes,
+                      size_t sizeBytes)
 {
     stateTransitionResourceBarrier(cmdList, srcBuffer, srcState, D3D12_RESOURCE_STATE_COPY_SOURCE);
     stateTransitionResourceBarrier(cmdList, destBuffer, destState, D3D12_RESOURCE_STATE_COPY_DEST);
